Renamer Fallible to FailsafeHandler, and moved it to util.. Renamed the test for it  FailsafeHandlerTest.
