+++ src/test/java/com/systemmeltdown/robotlib/util/FailsafeHandlerTest.java
@@ -1,16 +1,14 @@
-package com.systemmeltdown.robotlib.subsystems;
-
-import org.junit.Test;
+package com.systemmeltdown.robotlib.util;
 
 import edu.wpi.first.wpilibj.command.Subsystem;
-
+import org.junit.Test;
 import org.junit.Ignore;
 import static org.junit.Assert.*;
 
-public class SubsystemBaseTest {
+public class FailsafeHandlerTest {
 	@Test
 	public void testSetFailsafeIsActiveWhenSetToTrue() {
-		TestSubsystem testSub = new TestSubsystem();
+		TestFailsafeHandler testSub = new TestFailsafeHandler();
 
 		testSub.setFailsafeActive(true);
     
@@ -21,7 +19,7 @@ public class SubsystemBaseTest {
 
 	@Test
 	public void testSetFailsafeIsNotActiveWhenSetToFalse() {
-		TestSubsystem testSub = new TestSubsystem();
+		TestFailsafeHandler testSub = new TestFailsafeHandler();
 
 		testSub.setFailsafeActive(false);
     
@@ -32,7 +30,7 @@ public class SubsystemBaseTest {
 
 	@Test
 	public void testIsFailsafeIsNotActiveByDeefault() {
-		TestSubsystem testSub = new TestSubsystem();
+		TestFailsafeHandler testSub = new TestFailsafeHandler();
 
 		assertEquals( testSub.isFailsafeActive(), false );
 
@@ -40,7 +38,7 @@ public class SubsystemBaseTest {
 	}
 
 	@Ignore
-	public static class TestSubsystem extends Subsystem implements Fallible {
+	public static class TestFailsafeHandler extends Subsystem implements FailsafeHandler {
 		private boolean failsafeActive = false;
 		@Override
 		public void initDefaultCommand(){
